+++ lib/features/pet/domain/ui/utils.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+// --- App Constants ---
+
+class AppLayout {
+  static const double horizontalPadding = 20.0;
+  static const double sectionSpacing = 28.0;
+  static const double elementSpacing = 12.0;
+  static const double cardRadius = 16.0;
+  static const double minTouchTarget = 48.0;
+}
+
+class PetColors {
+  static const Color border = Color(0xFFE0E0E0);
+  static const Color cardBackground = Colors.white;
+  static const Color primary = Color(0xFF6200EE);
+  static const Color lightGrey = Color(0xFFF5F5F5);
+  static const Color textSecondary = Color(0xFF757575);
+}
