+++ lib/features/pet/domain/widgets/profile_section_header.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+
+import '../ui/utils.dart';
+
+// 프로필 섹션의 헤더를 표시하는 위젯
+class ProfileSectionHeader extends StatelessWidget {
+  final String title;
+  final VoidCallback onManagePressed;
+
+  const ProfileSectionHeader({
+    super.key,
+    required this.title,
+    required this.onManagePressed,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(
+        horizontal: AppLayout.horizontalPadding,
+      ),
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          Text(
+            title,
+            style: Theme.of(
+              context,
+            ).textTheme.titleLarge?.copyWith(fontWeight: FontWeight.bold),
+          ),
+          TextButton(onPressed: onManagePressed, child: const Text('관리하기')),
+        ],
+      ),
+    );
+  }
+}
