+++ pubspec.yaml
@@ -44,6 +44,7 @@ dependencies:
   shared_preferences: ^2.5.3
   flutter_dotenv: ^6.0.0
   google_sign_in: ^7.2.0
+  image_picker: ^1.0.7
 
 dev_dependencies:
   flutter_test:
@@ -54,7 +55,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^5.0.0
+  # flutter_lints: ^5.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
